flask/docker/requirements.txt
pip==23.0.1
Flask==2.0.3
Werkzeug==2.0.3
psycopg2-binary==2.9.1
